refactor change folder from api to controller
